# [PostRector][BC Break] Remove CurrentFileProvider on UseAddingPostRector and UseNodesToAddCollector