--- a/rules/CodingStyle/ClassNameImport/ClassNameImportSkipVoter/UsesClassNameImportSkipVoter.php
+++ b/rules/CodingStyle/ClassNameImport/ClassNameImportSkipVoter/UsesClassNameImportSkipVoter.php
@@ -26,7 +26,7 @@ final readonly class UsesClassNameImportSkipVoter implements ClassNameImportSkip
 
     public function shouldSkip(File $file, FullyQualifiedObjectType $fullyQualifiedObjectType, Node $node): bool
     {
-        $useImportTypes = $this->useNodesToAddCollector->getUseImportTypesByNode();
+        $useImportTypes = $this->useNodesToAddCollector->getUseImportTypes();
 
         foreach ($useImportTypes as $useImportType) {
             if (! $useImportType->equals($fullyQualifiedObjectType) && $useImportType->areShortNamesEqual(