@TomasVotruba here based on discussion on https://github.com/rectorphp/rector-src/pull/6049#pullrequestreview-2142272639