@samsonasik 2024-06-27
Tested on rector-src code base, got 1 seconds faster, it because there everyting already imported, so nothing to append:

**Before** **36.474** seconds

```
➜  rector-src git:(main) time bin/rector process --clear-cache
 2230/2230 [▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓] 100%

                                                                                                                        
 [OK] Rector is done!                                                                                                   
                                                                                                                        
bin/rector process --clear-cache  255.91s user 10.41s system 730% cpu 36.474 total
```

**After** **35.850** seconds

```
➜  rector-src git:(remove-current-file-provider) time bin/rector process --clear-cache    
 2230/2230 [▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓] 100%
                                                                                                                        
 [OK] Rector is done!                                                                                                   
                                                                                                                        
bin/rector process --clear-cache  252.20s user 9.94s system 731% cpu 35.850 total
```

Basically, this PR remove append by index `filePath`, instead, on next file, it reset to empty data instead.