@TomasVotruba 2024-06-27
Seems there is no difference in performance, but method API is now unclear.
Don't worry about this, I'll deep dive and explore it myself :)