@samsonasik 2024-06-27
The no different performance is because there is no file diff, as everything already imported in rector-src, to test this exactly, the big project with not have auto imported everywhere can be usage for test.