@TomasVotruba 2024-06-27
The goal here is not the performance, but to remove this setter/getter service. 
I made this bad design choice and now it's unclear what file is where :)